requirements.txt
torch==2.0.0
transformers==4.33.0
numpy==1.24.0
tqdm==4.65.0
Pillow==10.0.0
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.8.0
